Split declarations and definitions
Declarations in include/troll.hpp and definitions in src/troll.cpp
Updated CMakeLists.txt and .gitignore accordingly

Closes #7
